Move bitcoin specific crypto functions to provider level
