+++ package.json
@@ -38,7 +38,6 @@
     "bech32": "^1.1.3",
     "bs58": "^4.0.1",
     "crypto-js": "^3.1.9-1",
-    "fs": "^0.0.1-security",
     "json-bigint": "^0.3.0",
     "lodash": "^4.0.0",
     "regenerator-runtime": "^0.12.0"
