+++ src/providers/index.js
@@ -1,5 +1,6 @@
 import BitcoinRPCProvider from './bitcoin/BitcoinRPCProvider'
 import BitcoinLedgerProvider from './bitcoin/BitcoinLedgerProvider'
+import BitcoinCrypto from './bitcoin/BitcoinCrypto'
 
 import EthereumRPCProvider from './ethereum/EthereumRPCProvider'
 import EthereumLedgerProvider from './ethereum/EthereumLedgerProvider'
@@ -8,7 +9,8 @@ import EthereumMetaMaskProvider from './ethereum/EthereumMetaMaskProvider'
 export default {
   bitcoin: {
     BitcoinRPCProvider,
-    BitcoinLedgerProvider
+    BitcoinLedgerProvider,
+    crypto: BitcoinCrypto
   },
   ethereum: {
     EthereumRPCProvider,
